+++ repos/zimeg/CHANGELOG.md
@@ -8,6 +8,7 @@ and are noted with a `!` in the message.
 
 ## Changes
 
+- chore(pkgs/etime): bump version to 1.1.1 for fixed bugs with 1.21.6 2024-09-09
 - chore(pkgs/proximity-nvim): mark the latest tagged version of 0.1.0 2024-08-04
 - chore(pkgs/etime): bump version to 1.1.0 which introduces man pages 2024-08-03
 - chore(pkgs/etime): bump version from 1.0.1 to the latest 1.0.2 tags 2024-07-14